% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_phase_get.R
\name{log_phase_get}
\alias{log_phase_get}
\title{Get log phase}
\usage{
log_phase_get()
}
\value{
Character string with the name of the current log phase.
}
\description{
Get the current log phase.
}
\examples{
  path <- tempfile()
  log_phase_get()
  log_print(path = path)
  log_phase_set("different")
  log_phase_get()
  log_print(path = path)
  log_phase_reset()
  log_phase_get()
  log_read(path)
}
