% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_phase_reset.R
\name{log_phase_reset}
\alias{log_phase_reset}
\title{Reset log phase}
\usage{
log_phase_reset()
}
\value{
\code{NULL} (invisibly). Called for its side effects.
}
\description{
Reset the current log phase to the default value.
}
\examples{
  log_phase_get()
  log_phase_set("different")
  log_phase_get()
  log_phase_reset()
  log_phase_get()
}
