% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_version.R
\name{log_version}
\alias{log_version}
\title{Check the package version in the C code.}
\usage{
log_version()
}
\value{
Character string with the package version.
}
\description{
Not for users.
}
\examples{
  log_version()
}
\keyword{internal}
