% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_stat.R
\name{dir_stat}
\alias{dir_stat}
\title{Efficiently describe files in a directory.}
\usage{
dir_stat(
  path,
  units_size = c("megabytes", "bytes", "kilobytes", "gigabytes"),
  units_mtime = c("POSIXct", "numeric"),
  recent = NULL
)
}
\arguments{
\item{path}{Character string, file path to the directory of files
to describe.}

\item{units_size}{Character string with the units of the returned
\code{size} column in the output: \code{"megabytes"}, \code{"bytes"}, \code{"kilobytes"},
or \code{"gigabytes"}.}

\item{units_mtime}{Character string with the units of the returned
\code{mtime} column in the output with file modification time stamps.
Choices are \verb{"POSIXct} for a \code{POSIXct} time object or \code{"numeric"}
for an ordinary numeric vector.}

\item{recent}{Either \code{NULL} or an optional \code{"difftime"} object.
If a \code{"difftime"} object is supplied, then \code{\link[=dir_stat]{dir_stat()}}
only shows the most recently modified files in that time window.
For example, \code{recent = as.difftime(1.5, units = "hours")} tells
\code{\link[=dir_stat]{dir_stat()}} to only return information on files modified within
the last 1.5 hours.}
}
\value{
A data frame with one row per file and columns for the file path,
numeric size, and modification time stamp of each file.
The units of these last two columns are controlled by the
\code{units_size} and \code{units_mtime} arguments, respectively.
}
\description{
List the paths, sizes, modification times,
of all the regular files at the top level in a directory.
}
\details{
In large computational pipelines, it is common to end up with
tens of thousands of log files in a directory.
At this level of scale, \code{\link[base:file.info]{base::file.info()}}
is slow on older file systems.
\code{\link[=dir_stat]{dir_stat()}} can be up to 40 times faster.

\code{\link[=dir_stat]{dir_stat()}} is not recursive: it only queries regular files at the
top level of a directory. In addition, it follows symbolic links:
if a file is a link, then \code{\link[=dir_stat]{dir_stat()}} describes the file it points
to, rather than the link itself.
}
\examples{
  file.create(tempfile())
  file.create(tempfile())
  dir_stat(tempdir(), recent = as.difftime(1, units = "secs"))
}
\concept{utilities}
