% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_stat.R
\name{dir_stat}
\alias{dir_stat}
\title{Efficiently describe files in a directory.}
\usage{
dir_stat(path)
}
\arguments{
\item{path}{Character string, file path to the directory of files
to describe.}
}
\value{
A data frame with one row per file and columns for the file path,
size in bytes, and modification time stamp of each file.
}
\description{
List the paths, sizes, modification times,
of all the regular files at the top level in a directory.
}
\details{
In large computational pipelines, it is common to end up with
tens of thousands of log files in a directory.
At this level of scale, \code{\link[base:file.info]{base::file.info()}}
is slow on older file systems.
\code{\link[=dir_stat]{dir_stat()}} can be up to 40 times faster.

\code{\link[=dir_stat]{dir_stat()}} is not recursive: it only queries regular files at the
top level of a directory. In addition, it follows symbolic links:
if a file is a link, then \code{\link[=dir_stat]{dir_stat()}} describes the file it points
to, rather than the link itself.
}
\examples{
  file.create(tempfile())
  file.create(tempfile())
  dir_stat(tempdir())
}
\concept{utilities}
