% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_support.R
\name{log_support}
\alias{log_support}
\title{Log support}
\usage{
log_support()
}
\value{
\code{TRUE} if your system supports background logging, \code{FALSE}
otherwise.
}
\description{
Check if your system supports background logging.
}
\details{
The background logging functionality requires a Linux, Mac,
or Windows computer, It also requires POSIX thread support
and the \code{nanosleep()} C function.
}
\examples{
  log_support()
}
\seealso{
Other log: 
\code{\link{log_active}()},
\code{\link{log_plot}()},
\code{\link{log_print}()},
\code{\link{log_read}()},
\code{\link{log_start}()},
\code{\link{log_stop}()}
}
\concept{log}
